# openapi/api_delete_image_by_id.go
/*
 * Image Store
 *
 * Image store service provides Albums and Image management operaions
 *
 * API version: 1.0.0
 * Contact: diksha.may14@outlook.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"github.com/gorilla/mux"
	"net/http"
)

// A DeleteImageByIdApiController binds http requests to an api service and writes the service results to the http response
type DeleteImageByIdApiController struct {
	service DeleteImageByIdApiServicer
}

// NewDeleteImageByIdApiController creates a default api controller
func NewDeleteImageByIdApiController(s DeleteImageByIdApiServicer) Router {
	return &DeleteImageByIdApiController{service: s}
}

// Routes returns all of the api route for the DeleteImageByIdApiController
func (c *DeleteImageByIdApiController) Routes() Routes {
	return Routes{
		{
			"DeleteImageById",
			http.MethodDelete,
			"/image/{id}",
			c.DeleteImageById,
		},
	}
}

// DeleteImageById - delete an image with the the given image id
func (c *DeleteImageByIdApiController) DeleteImageById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	id := params["id"]
	result, err := c.service.DeleteImageById(id)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	EncodeJSONResponse(result, http.StatusNoContent, w)
}