# openapi/api_delete_images_by_album.go
/*
 * Image Store
 *
 * Image store service provides Albums and Image management operaions
 *
 * API version: 1.0.0
 * Contact: diksha.may14@outlook.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"github.com/gorilla/mux"
	"net/http"
)

// A DeleteImagesByAlbumApiController binds http requests to an api service and writes the service results to the http response
type DeleteImagesByAlbumApiController struct {
	service DeleteImagesByAlbumApiServicer
}

// NewDeleteImagesByAlbumApiController creates a default api controller
func NewDeleteImagesByAlbumApiController(s DeleteImagesByAlbumApiServicer) Router {
	return &DeleteImagesByAlbumApiController{service: s}
}

// Routes returns all of the api route for the DeleteImagesByAlbumApiController
func (c *DeleteImagesByAlbumApiController) Routes() Routes {
	return Routes{
		{
			"DeleteImagesByAlbum",
			http.MethodDelete,
			"/albums/{id}/images",
			c.DeleteImagesByAlbum,
		},
	}
}

// DeleteImagesByAlbum - delete all images in an album with the the given id
func (c *DeleteImagesByAlbumApiController) DeleteImagesByAlbum(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	id := params["id"]
	result, err := c.service.DeleteImagesByAlbum(id)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	EncodeJSONResponse(result, http.StatusNoContent, w)
}