# openapi/api_get_albums.go
/*
 * Image Store
 *
 * Image store service provides Albums and Image management operaions
 *
 * API version: 1.0.0
 * Contact: diksha.may14@outlook.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"image-store-service/constants"
	"net/http"
	"strings"
)

// A GetAlbumsApiController binds http requests to an api service and writes the service results to the http response
type GetAlbumsApiController struct {
	service GetAlbumsApiServicer
}

// NewGetAlbumsApiController creates a default api controller
func NewGetAlbumsApiController(s GetAlbumsApiServicer) Router {
	return &GetAlbumsApiController{service: s}
}

// Routes returns all of the api route for the GetAlbumsApiController
func (c *GetAlbumsApiController) Routes() Routes {
	return Routes{
		{
			"GetAlbums",
			strings.ToUpper("Get"),
			"/albums",
			c.GetAlbums,
		},
	}
}

// GetAlbums - Get all albums
func (c *GetAlbumsApiController) GetAlbums(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetAlbums()
	if err != nil {
		if strings.Contains(err.Error(), constants.ErrorDBRecordNotFound) || strings.Contains(err.Error(), constants.ErrorDBNoSuchTable) {
			w.WriteHeader(http.StatusNotFound)
			return
		}
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	EncodeJSONResponse(result, http.StatusOK, w)
}