# openapi/api_get_image_by_id.go
/*
 * Image Store
 *
 * Image store service provides Albums and Image management operaions
 *
 * API version: 1.0.0
 * Contact: diksha.may14@outlook.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"image-store-service/constants"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A GetImageByIdApiController binds http requests to an api service and writes the service results to the http response
type GetImageByIdApiController struct {
	service GetImageByIdApiServicer
}

// NewGetImageByIdApiController creates a default api controller
func NewGetImageByIdApiController(s GetImageByIdApiServicer) Router {
	return &GetImageByIdApiController{service: s}
}

// Routes returns all of the api route for the GetImageByIdApiController
func (c *GetImageByIdApiController) Routes() Routes {
	return Routes{
		{
			"GetImageById",
			strings.ToUpper("Get"),
			"/image/{id}",
			c.GetImageById,
		},
	}
}

// GetImageById - Get an image with the given image id
func (c *GetImageByIdApiController) GetImageById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	id := params["id"]
	result, err := c.service.GetImageById(id)
	if err != nil {
		if strings.Contains(err.Error(), constants.ErrorDBRecordNotFound) || strings.Contains(err.Error(), constants.ErrorDBNoSuchTable) {
			w.WriteHeader(http.StatusNotFound)
			return
		}
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	EncodeJSONResponse(result, http.StatusOK, w)
}