# openapi/model_response.go
/*
 * Image Store
 *
 * Image store service provides Albums and Image management operaions
 *
 * API version: 1.0.0
 * Contact: diksha.may14@outlook.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Response struct {
	Code string `json:"code,omitempty"`

	Message string `json:"message,omitempty"`
}